requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via flask
colorama==0.4.5
    # via click
flake8==5.0.4
    # via -r requirements.in
flask==2.2.2
    # via -r requirements.in
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
mccabe==0.7.0
    # via flake8
pycodestyle==2.9.1
    # via flake8
pyflakes==2.5.0
    # via flake8
python-dotenv==0.21.0
    # via -r requirements.in
requests==2.28.1
    # via -r requirements.in
ua-parser==0.16.1
    # via user-agents
urllib3==1.26.12
    # via requests
user-agents==2.2.0
    # via -r requirements.in
waitress==2.1.2
    # via -r requirements.in
werkzeug==2.2.2
    # via flask
